Combine/Merge/Group zimwiki unit test files
